ints to longs

+++ Meraki.Api/Data/SecureConnectNetworkDevice.cs
@@ -11,14 +11,14 @@ public class SecureConnectNetworkDevice
 	/// </summary>
 	[ApiAccess(ApiAccess.Read)]
 	[DataMember(Name = "organizationId")]
-	public int OrganizationId { get; set; }
+	public long OrganizationId { get; set; }
 
 	/// <summary>
 	/// The unique global identifier for this traffic source (origin). Use the origin ID to manage the device.The origin ID is not used by the device.
 	/// </summary>
 	[ApiAccess(ApiAccess.ReadUpdate)]
 	[DataMember(Name = "originId")]
-	public int OriginId { get; set; }
+	public long OriginId { get; set; }
 
 	/// <summary>
 	/// The time when the device was created. Specify an ISO 8601-formatted timestamp.
